fix display of mixed value for orientation
workaround for Qt QVariant bug where enums are interpreted as QWidget*

+++ src/entity/RPolylineEntity.cpp
@@ -296,7 +296,7 @@ QPair<QVariant, RPropertyAttributes> RPolylineEntity::getProperty(
 //                return qMakePair(v, attr);
 //            }
             QVariant v;
-            v.setValue(ori);
+            v.setValue((int)ori);
             return qMakePair(v, attr);
         }
